Move metrics stuff into its own module
Moved metrics stuff into its own module, which simplifies ClassificationModel
Remove some unused code.
